+++ include/internal/sockets.h
@@ -159,8 +159,8 @@ struct servent *PASCAL getservbyname(const char *, const char *);
 #  define openssl_fdset(a,b) FD_SET(a, b)
 # endif
 
-int BIO_socket_wait(int fd, int for_read, int timeout);
-int BIO_wait(BIO *bio, int timeout);
-int BIO_connect_retry(BIO *bio, int timeout);
+int BIO_socket_wait(int fd, int for_read, long timeout);
+int BIO_wait(BIO *bio, long timeout);
+int BIO_connect_retry(BIO *bio, long timeout);
 
 #endif
